Tests/Unit: use PSR-4
* Move the existing unit tests to a `Unit` subdirectory and ensure subdirectory names use "ProperCase".
* Remove prefixes from the test class names which are not needed in namespaced classes (`Yoast_WooCommerce`) and where relevant, update references to those classes in the test files.
* Ensure all test class file names match the name of the test class.
* Change the Composer `autoload-dev` directive to use PSR-4 autoloading.
* Update the `phpunit.xml.dist` file to match.
* Update the `phpcs.xml.dist` file to match.

Includes fixing copy/paste/typo errors in the test class docblocks and removing a redundant `@package` tag in a class docblock.
