+++ .phpcs.xml.dist
@@ -70,6 +70,12 @@
 				 path-to-namespace translations. -->
 			<property name="src_directory" type="array">
 				<element value="config"/>
+				<element value="tests/Unit"/>
+			</property>
+
+			<!-- Allow for the tests being in a non-standard directory (can be removed for YoastCS 3.0). -->
+			<property name="prefixes" type="array" extend="true">
+				<element value="Yoast\WP\Woocommerce\Tests\Unit"/>
 			</property>
 		</properties>
 	</rule>
@@ -103,7 +109,7 @@
 		<properties>
 			<property name="doubles_path" type="array">
 				<element value="/integration-tests/doubles"/>
-				<element value="/tests/doubles"/>
+				<element value="/tests/Unit/Doubles"/>
 			</property>
 		</properties>
 	</rule>
@@ -129,7 +135,7 @@
 	<!-- Allow for the double/mock classes to override methods just to change the visibility. -->
 	<rule ref="Generic.CodeAnalysis.UselessOverridingMethod.Found">
 		<exclude-pattern>/integration-tests/doubles/*\.php$</exclude-pattern>
-		<exclude-pattern>/tests/doubles/*\.php$</exclude-pattern>
+		<exclude-pattern>/tests/Unit/Doubles/*\.php$</exclude-pattern>
 	</rule>
 
 	<!-- Allow for select unit test files to use less secure code. -->
@@ -137,16 +143,10 @@
 		<exclude-pattern>/integration-tests/bootstrap\.php$</exclude-pattern>
 		<exclude-pattern>/integration-tests/framework/woocommerce-unittestcase\.php$</exclude-pattern>
 	</rule>
-
-	<!-- Disable filename sniff for TestCase.php for consistency with wordpress-seo -->
-	<rule ref="Yoast.Files.FileName.InvalidClassFileName">
-		<exclude-pattern>/tests/TestCase\.php$</exclude-pattern>
-	</rule>
-
 	<!-- Valid usage: For testing purposes, some non-prefixed globals are being created. -->
 	<rule ref="WordPress.NamingConventions.PrefixAllGlobals">
 		<exclude-pattern>/integration-tests/bootstrap\.php$</exclude-pattern>
-		<exclude-pattern>/tests/bootstrap\.php$</exclude-pattern>
+		<exclude-pattern>/tests/Unit/bootstrap\.php$</exclude-pattern>
 	</rule>
 
 	<!--
@@ -169,7 +169,15 @@
 
 	<!-- Disable max classname sniff in the tests/classes/presenters for now. -->
 	<rule ref="Yoast.NamingConventions.ObjectNameDepth.MaxExceeded">
-		<exclude-pattern>/tests/classes/presenters/*\.php$</exclude-pattern>
+		<exclude-pattern>/tests/Unit/Classes/Presenters/*\.php$</exclude-pattern>
+	</rule>
+
+	<!-- The below two issues will likely be fixed in YoastCS 3.0, after which these excludes can be removed. -->
+	<rule ref="Yoast.Files.FileName.InvalidClassFileName">
+		<exclude-pattern>/tests/Unit/*\.php$</exclude-pattern>
+	</rule>
+	<rule ref="Yoast.Commenting.FileComment.Unnecessary">
+		<exclude-pattern>/tests/Unit/bootstrap\.php$</exclude-pattern>
 	</rule>
 
 </ruleset>
