+++ phpunit.xml.dist
@@ -4,7 +4,7 @@
 	xsi:noNamespaceSchemaLocation="https://schema.phpunit.de/8.5/phpunit.xsd"
 		backupGlobals="false"
 		backupStaticAttributes="false"
-		bootstrap="tests/bootstrap.php"
+		bootstrap="tests/Unit/bootstrap.php"
 		colors="true"
 		convertDeprecationsToExceptions="true"
 		convertErrorsToExceptions="true"
@@ -20,7 +20,7 @@
 	>
 	<testsuites>
 		<testsuite name="wpseo-woocommerce">
-			<directory suffix="-test.php">./tests/</directory>
+			<directory suffix="Test.php">./tests/Unit/</directory>
 		</testsuite>
 	</testsuites>
