+++ tests/Unit/Classes/Presenters/Product_Price_Currency_Presenter_Test.php
@@ -1,13 +1,13 @@
 <?php
 
-namespace Yoast\WP\Woocommerce\Tests\Classes\Presenters;
+namespace Yoast\WP\Woocommerce\Tests\Unit\Classes\Presenters;
 
 use Brain\Monkey\Functions;
 use Mockery;
 use WC_Product;
 use WPSEO_WooCommerce_Product_Price_Currency_Presenter;
 use Yoast\WP\SEO\Presenters\Abstract_Indexable_Tag_Presenter;
-use Yoast\WP\Woocommerce\Tests\TestCase;
+use Yoast\WP\Woocommerce\Tests\Unit\TestCase;
 
 /**
  * Class Product_Price_Currency_Presenter_Test.
