+++ tests/Unit/Classes/Twitter_Test.php
@@ -1,12 +1,12 @@
 <?php
 
-namespace Yoast\WP\Woocommerce\Tests\Classes;
+namespace Yoast\WP\Woocommerce\Tests\Unit\Classes;
 
 use Brain\Monkey;
 use Mockery;
 use WPSEO_WooCommerce_Twitter;
 use Yoast\WP\SEO\Presentations\Indexable_Presentation;
-use Yoast\WP\Woocommerce\Tests\TestCase;
+use Yoast\WP\Woocommerce\Tests\Unit\TestCase;
 
 /**
  * Class Twitter_Test
